+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Tooltip: avoid removing the anchor `aria-describedby`
 -   Popover: unstable restore focus and changing anchor refs (triggered by anchor wrapped in tooltip)
 
+### Changed
+
+-   Autocomplete: forward `focusAnchorOnClose` prop
+
 ## [3.6.4][] - 2024-02-20
 
 ### Fixed